refactor(hmpb): move HMPB strings into `hmpb`
Rather than keeping them in `backend` and using them directly, pass them in when needed for generating audio. This ensures that `hmpb` will not try to import `backend`, which results in all kinds of problems when loading in the browser.
