+++ libs/backend/src/language_and_audio/index.ts
@@ -4,7 +4,6 @@ export * from './ballot_strings';
 export * from './audio';
 export * from './election_strings';
 export * from './election_package_strings';
-export * from './hmpb_strings';
 export * from './speech_synthesizer';
 export * from './test_utils';
 export * from './translator';
