+++ libs/hmpb/package.json
@@ -27,7 +27,6 @@
     "type-check": "tsc --build"
   },
   "dependencies": {
-    "@votingworks/backend": "workspace:*",
     "@votingworks/ballot-encoder": "workspace:*",
     "@votingworks/basics": "workspace:*",
     "@votingworks/fixtures": "workspace:*",
