+++ libs/hmpb/tsconfig.json
@@ -19,7 +19,6 @@
   },
   "references": [
     { "path": "../basics/tsconfig.build.json" },
-    { "path": "../backend/tsconfig.build.json" },
     { "path": "../ballot-encoder/tsconfig.build.json" },
     { "path": "../eslint-plugin-vx/tsconfig.build.json" },
     { "path": "../fixtures/tsconfig.build.json" },
